% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedforward.R
\name{addLayer}
\alias{addLayer}
\title{This function adds a new layer to the neural network 'net'
\code{addLayer} adds a new layer to the neural network}
\usage{
addLayer(net, input, neurons, activation)
}
\description{
@param net is the model to add a new layer.
 @param input number of inputs received by the neural network.
 @param neurons number of neurons in this layer.
 @param activation type of activation function, it can be "tansig", "logsig", "relu", "purelim", "softmax", "softplus".
}
