% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedforward.R
\name{sequencial_model}
\alias{sequencial_model}
\title{Retrieve a Java instance of SpellCorrector.}
\usage{
sequencial_model()
}
\value{
a Java instance of SpellCorrector
}
\description{
Retrieve a Java instance of SpellCorrector, with the training file
specified. Language model is trained before the instance is returned.
The spell corrector is adapted from Peter Norvig's demonstration.
}
